Create single tab file for all tabs
