--- ActionViewer/Tabs/NoEssenceTab.cs
@@ -1,24 +1,0 @@
-﻿using ActionViewer.Functions;
-using ActionViewer.Windows;
-using Dalamud.Game.ClientState.Objects.SubKinds;
-using ImGuiNET;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace ActionViewer.Tabs;
-
-public class NoEssenceTab : MainWindowTab
-{
-    private string searchText = string.Empty;
-
-    public NoEssenceTab(Plugin plugin) : base("No Ess.", plugin) { }
-
-    public override void Draw()
-    {
-        List<IPlayerCharacter> playerCharacters = this.Plugin.ActionViewer.getPlayerCharacters();
-
-        ImGui.SetNextItemWidth(-1 * ImGui.GetIO().FontGlobalScale);
-        ImGui.InputText("", ref searchText, 256);
-        StatusInfoFunctions.GenerateStatusTable(playerCharacters, searchText, this.Plugin.Configuration.AnonymousMode, this.Plugin.Configuration.Tooltips, this.Plugin.Configuration.TargetRangeLimit, this.Plugin.ActionSheet, this.Plugin.ItemSheet, "noEss");
-    }
-}
